Default platform is now specified
Perhaps this should be generalized to other functions

From the doc:

3: How will the ICD handle a NULL cl_platform_id?
RESOLVED: The ICD will by default choose the first enumerated platform as the NULL
platform. The user can override this default by setting an environment variable
OPENCL_ICD_DEFAULT_PLATFORM to the desired platform index. The API calls that deal
with platforms will return CL_INVALID_PLATFORM if the index is not between zero and
(number of platforms - 1), both inclusive.