Fallback to dispatch table for clGetPlatformInfo
Currently, we call clGetPlatformInfo by directly loading the function from the
vendor library. Although this is suggested by the ICD spec, many vendors do not
directly expose clGetPlatformInfo, which breaks ICD loading on such libraries.

Fix this by calling clGetPlatformInfo using the dispatch table when
clGetPlatformInfo is not exported.  This is a looser interpretation of the spec,
but it unbreaks many vendor libraries.